Split ChangePassphraseModal into logic and presentation components
